aidemy-bootstrap/portal/requirements.txt
google-genai==1.0.0
flask==3.1.0
jsonify==0.5
langchain_google_vertexai==2.0.13
langchain_core==0.3.34
pydantic==2.10.5
google-cloud-storage==2.19.0
